async_api/tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
aioelasticsearch==0.7.0
aioredis==1.3.1
pydantic